requirements.txt
opencv-python
numpy
pytesseract
scipy
scikit-learn
python-dotenv